Seperation of concerns in cssAggregator for testability.
